# docs/api/backend-v1.openapi.yaml
openapi: 3.0.3
info:
  title: Backend API
  description: |
    REST v1: бот (диалог, сдачи), структура потока и **веб-клиент** (дашборд преподавателя,
    лидерборд, ленты, матрица прогресса, веб-сессия и диалог по `participant_id`).
    Канонический контракт репозитория; реализация отдельных путей — см. tasklist backend / frontend.

    **Сдачи:** `POST /v1/submissions` — создание одной сдачи (бот). `GET /v1/flows/{flow_id}/submissions` —
    лента сдач по потоку (веб/аналитика). Это разные ресурсы; путать методы и пути не следует.

    **Голос:** `POST /v1/voice/dialog-messages` — multipart аудио, STT Whisper на backend (см. ADR adr-004).

    **Запрос к данным потока:** `POST /v1/flows/{flow_id}/data-query` — только роль преподавателя; классификация вопроса через LLM,
    выполнение — фиксированные SELECT (см. ADR adr-005).

    **Политика 404 и 403:** для защищённых ресурсов по UUID сервер может возвращать `403`, а не `404`,
    чтобы не раскрывать факт существования объекта — конкретное поведение фиксируется в реализации.

    **Несовместимые изменения v1:** удаление полей, смена типа или смысла поля, изменение кодов ошибок
    или формы `ErrorBody` — только с новым префиксом версии или явной миграцией клиентов.

    Ответы `429` и `500`: указаны явно на чувствительных операциях; остальные методы при сбое также могут
    вернуть `500` с телом `ErrorBody` (`code: internal_error`).
  version: 1.0.0
servers:
  - url: /
...
paths:
  /v1/dialog-messages:
    post:
      tags: [dialog-messages]
      summary: Отправить сообщение и получить ответ ассистента
      description: |
        Резолв участника по `telegram_user_id` и `flow_id`, сохранение сообщений
        (`DialogMessage` с ролями user/assistant), ответ с текстом от LLM (логика в итерации 5).
      operationId: createDialogMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DialogMessageCreateRequest'
      responses:
        '201':
          description: Сообщения сохранены, ответ ассистента сформирован
          headers:
            Location:
              description: |
                Коллекция сообщений участника в потоке; идентификатор новой user-реплики — в теле (`user_message_id`).
              schema:
                type: string
                example: /v1/dialog-messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogMessageCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/voice/dialog-messages:
    post:
      tags: [dialog-messages]
      summary: Голосовое сообщение в диалог (STT + ответ ассистента)
      description: |
        Multipart: поле `audio` (binary), форма `flow_id` (UUID).
        При **Bearer = INTERNAL_API_TOKEN** обязательно поле формы `telegram_user_id`
        (как у `POST /v1/dialog-messages`): резолв участника как у бота.
        При **Bearer = JWT веб-сессии** участник берётся из токена; `flow_id` должен совпадать с claim `flow_id`.
        В ответе поле `transcription` — распознанный текст (дублируется как содержимое user-сообщения в БД).
      operationId: createVoiceDialogMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio, flow_id]
              properties:
                audio:
                  type: string
                  format: binary
                  description: Аудио (webm, ogg/opus и др., в пределах лимита сервера)
                flow_id:
                  type: string
                  format: uuid
                telegram_user_id:
                  type: integer
                  format: int64
                  description: Только для вызова с internal token (бот)
      responses:
        '201':
          description: Транскрипция сохранена как user-сообщение, ответ ассистента сформирован
          headers:
            Location:
              description: Коллекция диалога участника (веб URI при JWT; общий паттерн — как у текстового POST)
              schema:
                type: string
                example: /v1/dialog-messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogMessageCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: Файл слишком большой
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: STT недоступен (нет OPENAI_API_KEY или сбой провайдера)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/submissions:
    post:
      tags: [submissions]
      summary: Зафиксировать сдачу задания
      description: |
        Создаёт `Submission` со статусом `submitted`. Задание должно относиться к тому же потоку,
        что и участник (через цепочку Lesson → Module → Flow).
      operationId: createSubmission
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionCreateRequest'
      responses:
        '201':
          description: Сдача создана
          headers:
            Location:
              description: URI созданной сдачи (тот же идентификатор, что `id` в теле)
              schema:
                type: string
                example: /v1/submissions/550e8400-e29b-41d4-a716-446655440000
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/participants/{participant_id}/submissions:
    get:
      tags: [participants]
      summary: Сдачи участника
      description: |
        Все сдачи участника, новые первыми. MVP: ответ — массив без обёртки; опционально `limit` усечёт выдачу на сервере.
        При большом числе записей планируется та же пагинация `limit`/`cursor`, что у `GET /v1/flows/{flow_id}/submissions`.
      operationId: listParticipantSubmissions
      security:
        - bearerAuth: []
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Необязательный лимит записей (MVP; при отсутствии — поведение сервера по умолчанию)
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Список сдач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParticipantSubmissionRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/flows/{flow_id}/submissions:
    get:
      tags: [flows]
      summary: Лента сдач по потоку
      description: Новые первыми. Для преподавателя; элементы могут включать материалы урока.
      operationId: listFlowSubmissionFeed
      security:
        - bearerAuth: []
      parameters:
        - name: flow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: Страница ленты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionFeedPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/participants/{participant_id}/dialog-messages:
    get:
      tags: [dialog-messages]
      summary: История диалога участника (веб)
      operationId: listParticipantDialogMessages
      security:
        - bearerAuth: []
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: flow_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: Страница сообщений
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogMessageListPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [dialog-messages]
      summary: Отправить сообщение в диалог (веб)
      description: Создаёт сообщение пользователя и ответ ассистента; проверка прав по `flow_id` и `participant_id`.
      operationId: createParticipantDialogMessage
      security:
        - bearerAuth: []
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: flow_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantDialogMessageCreateRequest'
      responses:
        '201':
          description: Сообщения сохранены, ответ ассистента сформирован
          headers:
            Location:
              description: Коллекция диалога участника; новые id — в теле ответа.
              schema:
                type: string
                example: /v1/participants/550e8400-e29b-41d4-a716-446655440000/dialog-messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogMessageCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `INTERNAL_API_TOKEN` (бот из `.env`) или JWT веб-сессии из `POST /v1/auth/web-session`.
        См. docs/integrations.md и docs/tasks/impl/frontend/iteration-0-ui-api-contracts/plan.md.
  schemas:
    DialogMessageCreateRequest:
      type: object
      required: [flow_id, telegram_user_id, content]
      properties:
        flow_id:
          type: string
          format: uuid
        telegram_user_id:
          type: integer
          format: int64
          example: 123456789
        content:
          type: string
          minLength: 1
          description: Текст сообщения пользователя (после trim непустая строка)
    DialogMessageCreateResponse:
      type: object
      required: [reply_text, user_message_id, assistant_message_id]
      properties:
        reply_text:
          type: string
        user_message_id:
          type: string
          format: uuid
        assistant_message_id:
          type: string
          format: uuid
        transcription:
          type: string
          nullable: true
          description: При голосовом заполнено распознанным текстом пользователя (опционально)
...
    CursorPageMeta:
      type: object
      required: [next_cursor]
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Пусто если страница последняя
...
    SubmissionFeedPage:
      allOf:
        - $ref: '#/components/schemas/CursorPageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SubmissionFeedItem'
    MaterialRef:
      type: object
      required: [id, title, type]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [link, file, text]
        url:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
    SubmissionFeedItem:
      type: object
      required:
        - submission_id
        - participant_id
        - participant_name
        - lesson_id
        - lesson_title
        - assignment_id
        - assignment_title
        - status
        - submitted_at
        - materials
      properties:
        submission_id:
          type: string
          format: uuid
        participant_id:
          type: string
          format: uuid
        participant_name:
          type: string
        lesson_id:
          type: string
          format: uuid
        lesson_title:
          type: string
        assignment_id:
          type: string
          format: uuid
        assignment_title:
          type: string
        status:
          type: string
          enum: [submitted, reviewed, approved]
        submitted_at:
          type: string
          format: date-time
        comment:
          type: string
          nullable: true
        materials:
          type: array
          items:
            $ref: '#/components/schemas/MaterialRef'
...
    DialogMessageRead:
      type: object
      required: [id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        created_at:
          type: string
          format: date-time
    DialogMessageListPage:
      allOf:
        - $ref: '#/components/schemas/CursorPageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/DialogMessageRead'
    ParticipantDialogMessageCreateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    SubmissionCreateRequest:
      type: object
      required: [flow_id, telegram_user_id, assignment_id]
      properties:
        flow_id:
          type: string
          format: uuid
        telegram_user_id:
          type: integer
          format: int64
        assignment_id:
          type: string
          format: uuid
        comment:
          type: string
          nullable: true
...
    ParticipantSubmissionRead:
      type: object
      required: [id, assignment_id, status, submitted_at]
      properties:
        id:
          type: string
          format: uuid
        assignment_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [submitted, reviewed, approved]
        submitted_at:
          type: string
          format: date-time
        comment:
          type: string
          nullable: true
    SubmissionResponse:
      type: object
      required: [id, assignment_id, participant_id, status, submitted_at]
      properties:
        id:
          type: string
          format: uuid
        assignment_id:
          type: string
          format: uuid
        participant_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [submitted, reviewed, approved]
        submitted_at:
          type: string
          format: date-time
          description: ISO 8601 в UTC
        comment:
          type: string
          nullable: true
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Стабильный код для ветвления на клиенте
            message:
              type: string
              description: Краткое описание; не использовать для логики клиента
            details:
              type: array
              nullable: true
              description: Опционально; ошибки валидации по полям (клиент ветвит по `code`, не по `message`)
              items:
                type: object
                required: [field, issue]
                properties:
                  field:
                    type: string
                    description: Имя поля или path (например body.telegram_username)
                  issue:
                    type: string
                    description: Краткое описание проблемы для разработчика
  responses:
    BadRequest:
      description: Невалидный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            validation:
              value:
                error:
                  code: validation_error
                  message: Invalid request body
                  details:
                    - field: telegram_username
                      issue: must not be empty
    Unauthorized:
      description: Нет или неверный Bearer-токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            default:
              value:
                error:
                  code: unauthorized
                  message: Authentication required
    Forbidden:
      description: Доступ запрещён (роль / политика потока)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            default:
              value:
                error:
                  code: forbidden
                  message: Operation not allowed for this participant
    NotFound:
      description: |
        Ресурс не найден или связь участник–поток отсутствует (см. стабильный `error.code` в теле).
        При доступе к чужим данным по UUID реализация может вернуть `403` вместо `404` — см. описание API в `info`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            flow:
              value:
                error:
                  code: flow_not_found
                  message: Flow not found
            participant:
              value:
                error:
                  code: participant_not_found
                  message: User is not a participant of this flow
            assignment:
              value:
                error:
                  code: assignment_not_found
                  message: Assignment not found or not in this flow
            user_not_found:
              value:
                error:
                  code: user_not_found
                  message: No user matches this telegram username
            user_not_in_flow:
              value:
                error:
                  code: user_not_in_flow
                  message: User is not a participant of the given flow
    Conflict:
      description: Конфликт бизнес-правил
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            duplicate:
              value:
                error:
                  code: submission_already_exists
                  message: Submission already exists for this assignment and participant
    TooManyRequests:
      description: Превышен лимит запросов (например к LLM); при наличии — заголовок Retry-After
      headers:
        Retry-After:
          description: Рекомендуемая пауза перед повтором, секунды
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            default:
              value:
                error:
                  code: rate_limited
                  message: Too many requests
...
    InternalServerError:
      description: Внутренняя ошибка; тело без stack trace и внутренних путей
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            default:
              value:
                error:
                  code: internal_error
                  message: An unexpected error occurred